// src/types/content/CardType.ts
/** Tipagem do conteúdo dos cards com links para as páginas. */
export interface CardLinkType {
  /** Título do card. */
  title: string
  /** Descrição no body do card. */
  description: string
  /** Link para a página do card. */
  href: string
  /** Texto do botão do card. */
  buttonText: string
  /** Imagem do card. */
  image: string
}